# FinalProject/docs/openapi.yaml
...
openapi: 3.1.0
info:
  title: Credit Default Risk Scoring API
  description: "\nOnline scoring service for **credit default risk** (UCI Credit Default,\
    \ 23 features).\n\n* `/api/v1/*` requires the `X-API-Key` header.\n* Every response\
    \ carries `X-Request-ID` (send your own to correlate logs; otherwise one is generated).\n\
    * Errors share one schema: `{code, message, details, request_id}`.\n* `/health/live`,\
    \ `/health/ready` and `/metrics` are unauthenticated for orchestrators and Prometheus;\n\
    \  keep them on the internal network.\n"
  version: 1.1.0
paths:
  /api/v1/predict:
    post:
      tags:
      - prediction
      summary: Score one cardholder
      description: Realtime decision for a cardholder request made in the mobile app
        / internet banking (credit limit increase, cash advance, instalment conversion).
        Returns next-cycle default probability, APPROVE/REVIEW/DECLINE decision, credit
        score, tier, recommended limit, rule-based risk factors and policy guardrails.
        The request is logged for drift analysis.
      operationId: predict_api_v1_predict_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPredictRequest'
            examples:
              low_risk:
                summary: Low-risk cardholder
                description: Pays in full every month, low utilization.
                value:
                  LIMIT_BAL: 200000.0
                  SEX: 2
                  EDUCATION: 1
                  MARRIAGE: 1
                  AGE: 38
                  PAY_0: -1
                  PAY_2: -1
                  PAY_3: -1
                  PAY_4: -1
                  PAY_5: -1
                  PAY_6: -1
                  BILL_AMT1: 12000.0
                  BILL_AMT2: 11500.0
                  BILL_AMT3: 11000.0
                  BILL_AMT4: 10500.0
                  BILL_AMT5: 10000.0
                  BILL_AMT6: 9500.0
                  PAY_AMT1: 12000.0
                  PAY_AMT2: 11500.0
                  PAY_AMT3: 11000.0
                  PAY_AMT4: 10500.0
                  PAY_AMT5: 10000.0
                  PAY_AMT6: 9500.0
              high_risk:
                summary: High-risk cardholder
                description: Two months late on every statement, card maxed out.
                value:
                  LIMIT_BAL: 20000.0
                  SEX: 1
                  EDUCATION: 2
                  MARRIAGE: 2
                  AGE: 23
                  PAY_0: 2
                  PAY_2: 2
                  PAY_3: 2
                  PAY_4: 2
                  PAY_5: 2
                  PAY_6: 2
                  BILL_AMT1: 19500.0
                  BILL_AMT2: 19800.0
                  BILL_AMT3: 19900.0
                  BILL_AMT4: 20000.0
                  BILL_AMT5: 20100.0
                  BILL_AMT6: 20200.0
                  PAY_AMT1: 0.0
                  PAY_AMT2: 0.0
                  PAY_AMT3: 0.0
                  PAY_AMT4: 0.0
                  PAY_AMT5: 0.0
                  PAY_AMT6: 0.0
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PredictionResponse'
        '401':
          description: API key missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MISSING_API_KEY
                message: Missing X-API-Key header.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '403':
          description: API key invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_API_KEY
                message: The provided API key is not valid.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '422':
          description: Payload failed validation (types, domain bounds, unknown fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: Request validation failed.
                details:
                - field: AGE
                  message: Input should be greater than or equal to 18
                  type: greater_than_equal
                  constraint:
                    ge: '18'
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: Internal server error.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '503':
          description: No model loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MODEL_UNAVAILABLE
                message: No model is loaded; check GET /health/ready.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
      security:
      - ApiKeyAuth: []
  /api/v1/predict/batch:
    post:
      tags:
      - prediction
      summary: Score several cardholders
      description: Vectorized scoring of up to `batch_max_size` cardholders (configs/serving.yaml,
        default 500), used by the post-statement limit review and early-warning jobs.
        Results keep request order; each item gets its own `<request_id>-<index>`
        id in the inference log. The request field is `cardholders`; the legacy `applicants`
        field is a deprecated alias (send one of them, never both).
      operationId: predict_batch_api_v1_predict_batch_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchPredictRequest'
            examples:
              two_cardholders:
                summary: Two cardholders
                description: Post-statement limit review of two accounts.
                value:
                  cardholders:
                  - LIMIT_BAL: 200000.0
                    SEX: 2
                    EDUCATION: 1
                    MARRIAGE: 1
                    AGE: 38
                    PAY_0: -1
                    PAY_2: -1
                    PAY_3: -1
                    PAY_4: -1
                    PAY_5: -1
                    PAY_6: -1
                    BILL_AMT1: 12000.0
                    BILL_AMT2: 11500.0
                    BILL_AMT3: 11000.0
                    BILL_AMT4: 10500.0
                    BILL_AMT5: 10000.0
                    BILL_AMT6: 9500.0
                    PAY_AMT1: 12000.0
                    PAY_AMT2: 11500.0
                    PAY_AMT3: 11000.0
                    PAY_AMT4: 10500.0
                    PAY_AMT5: 10000.0
                    PAY_AMT6: 9500.0
                  - LIMIT_BAL: 20000.0
                    SEX: 1
                    EDUCATION: 2
                    MARRIAGE: 2
                    AGE: 23
                    PAY_0: 2
                    PAY_2: 2
                    PAY_3: 2
                    PAY_4: 2
                    PAY_5: 2
                    PAY_6: 2
                    BILL_AMT1: 19500.0
                    BILL_AMT2: 19800.0
                    BILL_AMT3: 19900.0
                    BILL_AMT4: 20000.0
                    BILL_AMT5: 20100.0
                    BILL_AMT6: 20200.0
                    PAY_AMT1: 0.0
                    PAY_AMT2: 0.0
                    PAY_AMT3: 0.0
                    PAY_AMT4: 0.0
                    PAY_AMT5: 0.0
                    PAY_AMT6: 0.0
              deprecated_alias:
                summary: Deprecated `applicants` field
                description: Same request with the deprecated field name; returns
                  the same result. Prefer `cardholders`.
                value:
                  applicants:
                  - LIMIT_BAL: 200000.0
                    SEX: 2
                    EDUCATION: 1
                    MARRIAGE: 1
                    AGE: 38
                    PAY_0: -1
                    PAY_2: -1
                    PAY_3: -1
                    PAY_4: -1
                    PAY_5: -1
                    PAY_6: -1
                    BILL_AMT1: 12000.0
                    BILL_AMT2: 11500.0
                    BILL_AMT3: 11000.0
                    BILL_AMT4: 10500.0
                    BILL_AMT5: 10000.0
                    BILL_AMT6: 9500.0
                    PAY_AMT1: 12000.0
                    PAY_AMT2: 11500.0
                    PAY_AMT3: 11000.0
                    PAY_AMT4: 10500.0
                    PAY_AMT5: 10000.0
                    PAY_AMT6: 9500.0
                  - LIMIT_BAL: 20000.0
                    SEX: 1
                    EDUCATION: 2
                    MARRIAGE: 2
                    AGE: 23
                    PAY_0: 2
                    PAY_2: 2
                    PAY_3: 2
                    PAY_4: 2
                    PAY_5: 2
                    PAY_6: 2
                    BILL_AMT1: 19500.0
                    BILL_AMT2: 19800.0
                    BILL_AMT3: 19900.0
                    BILL_AMT4: 20000.0
                    BILL_AMT5: 20100.0
                    BILL_AMT6: 20200.0
                    PAY_AMT1: 0.0
                    PAY_AMT2: 0.0
                    PAY_AMT3: 0.0
                    PAY_AMT4: 0.0
                    PAY_AMT5: 0.0
                    PAY_AMT6: 0.0
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchPredictionResponse'
        '401':
          description: API key missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MISSING_API_KEY
                message: Missing X-API-Key header.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '403':
          description: API key invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_API_KEY
                message: The provided API key is not valid.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '413':
          description: Batch larger than the configured maximum
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: BATCH_TOO_LARGE
                message: Batch contains 800 cardholders; the maximum is 500.
                details:
                  max_size: 500
                  received: 800
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '422':
          description: Payload failed validation (types, domain bounds, unknown fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: Request validation failed.
                details:
                - field: AGE
                  message: Input should be greater than or equal to 18
                  type: greater_than_equal
                  constraint:
                    ge: '18'
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: Internal server error.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '503':
          description: No model loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MODEL_UNAVAILABLE
                message: No model is loaded; check GET /health/ready.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
      security:
      - ApiKeyAuth: []
  /api/v1/explain:
    post:
      tags:
      - prediction
      summary: Score one cardholder and explain the score
      description: 'Used by credit risk analysts for accounts in REVIEW/DECLINE and
        for adverse action notices. Adds local, model-based attributions (top-k by
        magnitude). Default method `shap_permutation`: SHAP values against the median
        training cardholder, so contributions sum to `default_probability - reference_probability`,
        and `top_risk_factors` is derived from them. If SHAP fails the endpoint degrades
        to `reference_substitution` (each feature replaced by the reference value).
        Explanations are not written to the inference log.'
      operationId: explain_api_v1_explain_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreditPredictRequest'
            examples:
              low_risk:
                summary: Low-risk cardholder
                description: Pays in full every month, low utilization.
                value:
                  LIMIT_BAL: 200000.0
                  SEX: 2
                  EDUCATION: 1
                  MARRIAGE: 1
                  AGE: 38
                  PAY_0: -1
                  PAY_2: -1
                  PAY_3: -1
                  PAY_4: -1
                  PAY_5: -1
                  PAY_6: -1
                  BILL_AMT1: 12000.0
                  BILL_AMT2: 11500.0
                  BILL_AMT3: 11000.0
                  BILL_AMT4: 10500.0
                  BILL_AMT5: 10000.0
                  BILL_AMT6: 9500.0
                  PAY_AMT1: 12000.0
                  PAY_AMT2: 11500.0
                  PAY_AMT3: 11000.0
                  PAY_AMT4: 10500.0
                  PAY_AMT5: 10000.0
                  PAY_AMT6: 9500.0
              high_risk:
                summary: High-risk cardholder
                description: Two months late on every statement, card maxed out.
                value:
                  LIMIT_BAL: 20000.0
                  SEX: 1
                  EDUCATION: 2
                  MARRIAGE: 2
                  AGE: 23
                  PAY_0: 2
                  PAY_2: 2
                  PAY_3: 2
                  PAY_4: 2
                  PAY_5: 2
                  PAY_6: 2
                  BILL_AMT1: 19500.0
                  BILL_AMT2: 19800.0
                  BILL_AMT3: 19900.0
                  BILL_AMT4: 20000.0
                  BILL_AMT5: 20100.0
                  BILL_AMT6: 20200.0
                  PAY_AMT1: 0.0
                  PAY_AMT2: 0.0
                  PAY_AMT3: 0.0
                  PAY_AMT4: 0.0
                  PAY_AMT5: 0.0
                  PAY_AMT6: 0.0
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExplainResponse'
        '401':
          description: API key missing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MISSING_API_KEY
                message: Missing X-API-Key header.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '403':
          description: API key invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INVALID_API_KEY
                message: The provided API key is not valid.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '422':
          description: Payload failed validation (types, domain bounds, unknown fields)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: VALIDATION_ERROR
                message: Request validation failed.
                details:
                - field: AGE
                  message: Input should be greater than or equal to 18
                  type: greater_than_equal
                  constraint:
                    ge: '18'
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '500':
          description: Unexpected server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: INTERNAL_ERROR
                message: Internal server error.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        '503':
          description: No model loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: MODEL_UNAVAILABLE
                message: No model is loaded; check GET /health/ready.
                request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
      security:
      - ApiKeyAuth: []
...
components:
  schemas:
    BatchPredictRequest:
      properties:
        cardholders:
          items:
            $ref: '#/components/schemas/CreditPredictRequest'
          type: array
          minItems: 1
          title: Cardholders
          description: Cardholder accounts to score; the server caps the size (see
            `batch_max_size`). Validation errors report the field name the client
            sent (`cardholders.N.FIELD` or `applicants.N.FIELD`).
        applicants:
          items:
            $ref: '#/components/schemas/CreditPredictRequest'
          type: array
          minItems: 1
          title: Applicants
          description: 'Deprecated alias of `cardholders`, still accepted for backward
            compatibility. Send exactly one of the two fields: a request carrying
            both is rejected with 422.'
          deprecated: true
      additionalProperties: false
      type: object
      required:
      - cardholders
      title: BatchPredictRequest
      description: Several cardholder accounts scored in one vectorized call (post-statement
        limit review).
...
    CreditPredictRequest:
      properties:
        LIMIT_BAL:
          type: number
          maximum: 10000000.0
          exclusiveMinimum: 0.0
          title: Limit Bal
          description: Current credit limit of the card (NT dollar)
        SEX:
          type: integer
          maximum: 2.0
          minimum: 1.0
          title: Sex
          description: Gender (1=male, 2=female)
        EDUCATION:
          type: integer
          maximum: 6.0
          minimum: 0.0
          title: Education
          description: Education (1=graduate, 2=university, 3=high school, 4=others,
            0/5/6=unknown)
        MARRIAGE:
          type: integer
          maximum: 3.0
          minimum: 0.0
          title: Marriage
          description: Marital status (1=married, 2=single, 3=others, 0=unknown)
        AGE:
          type: integer
          maximum: 100.0
          minimum: 18.0
          title: Age
          description: Age in years
        PAY_0:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 0
          description: Repayment status in September (-2=no consumption, -1=paid in
            full, 0=revolving credit, 1..9=months of payment delay)
        PAY_2:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 2
          description: Repayment status in August
        PAY_3:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 3
          description: Repayment status in July
        PAY_4:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 4
          description: Repayment status in June
        PAY_5:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 5
          description: Repayment status in May
        PAY_6:
          type: integer
          maximum: 9.0
          minimum: -2.0
          title: Pay 6
          description: Repayment status in April
        BILL_AMT1:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt1
          description: Bill statement in September (NT dollar, may be negative)
        BILL_AMT2:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt2
          description: Bill statement in August
        BILL_AMT3:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt3
          description: Bill statement in July
        BILL_AMT4:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt4
          description: Bill statement in June
        BILL_AMT5:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt5
          description: Bill statement in May
        BILL_AMT6:
          type: number
          maximum: 10000000.0
          minimum: -10000000.0
          title: Bill Amt6
          description: Bill statement in April
        PAY_AMT1:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt1
          description: Previous payment in September (NT dollar)
        PAY_AMT2:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt2
          description: Previous payment in August
        PAY_AMT3:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt3
          description: Previous payment in July
        PAY_AMT4:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt4
          description: Previous payment in June
        PAY_AMT5:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt5
          description: Previous payment in May
        PAY_AMT6:
          type: number
          maximum: 10000000.0
          minimum: 0.0
          title: Pay Amt6
          description: Previous payment in April
      additionalProperties: false
      type: object
      required:
      - LIMIT_BAL
      - SEX
      - EDUCATION
      - MARRIAGE
      - AGE
      - PAY_0
      - PAY_2
      - PAY_3
      - PAY_4
      - PAY_5
      - PAY_6
      - BILL_AMT1
      - BILL_AMT2
      - BILL_AMT3
      - BILL_AMT4
      - BILL_AMT5
      - BILL_AMT6
      - PAY_AMT1
      - PAY_AMT2
      - PAY_AMT3
      - PAY_AMT4
      - PAY_AMT5
      - PAY_AMT6
      title: CreditPredictRequest
      description: One cardholder account after its latest statement; field names
        follow the UCI dataset columns.
      examples:
      - AGE: 38
        BILL_AMT1: 12000.0
        BILL_AMT2: 11500.0
        BILL_AMT3: 11000.0
        BILL_AMT4: 10500.0
        BILL_AMT5: 10000.0
        BILL_AMT6: 9500.0
        EDUCATION: 1
        LIMIT_BAL: 200000.0
        MARRIAGE: 1
        PAY_0: -1
        PAY_2: -1
        PAY_3: -1
        PAY_4: -1
        PAY_5: -1
        PAY_6: -1
        PAY_AMT1: 12000.0
        PAY_AMT2: 11500.0
        PAY_AMT3: 11000.0
        PAY_AMT4: 10500.0
        PAY_AMT5: 10000.0
        PAY_AMT6: 9500.0
        SEX: 2
...
    ErrorResponse:
      properties:
        code:
          type: string
          title: Code
          description: Stable machine-readable error code
        message:
          type: string
          title: Message
        details:
          title: Details
        request_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Request Id
      type: object
      required:
      - code
      - message
      title: ErrorResponse
      description: Body of every non-2xx response.
      examples:
      - code: VALIDATION_ERROR
        details:
        - constraint:
            ge: '18'
          field: AGE
          message: Input should be greater than or equal to 18
          type: greater_than_equal
        message: Request validation failed.
        request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
...
    PredictionResponse:
      properties:
        model_version:
          type: string
          title: Model Version
          description: Registry version number, or the artifact name in fallback mode
        served_by:
          type: string
          title: Served By
          description: 'Model source: mlflow_registry or local_artifact'
        default_prediction:
          type: integer
          title: Default Prediction
          description: 0 = no default expected, 1 = default risk
        default_probability:
          type: number
          maximum: 1.0
          minimum: 0.0
          title: Default Probability
          description: Probability of default (0.0 to 1.0)
        credit_score:
          type: integer
          maximum: 850.0
          minimum: 300.0
          title: Credit Score
          description: Credit score on a 300-850 scale (higher is better)
        credit_tier:
          type: string
          enum:
          - PRIME
          - NEAR_PRIME
          - SUBPRIME
          - HIGH_RISK
          title: Credit Tier
          description: Risk tier
        risk_decision:
          type: string
          enum:
          - APPROVE
          - REVIEW
          - DECLINE
          title: Risk Decision
          description: Business decision
        recommended_limit_ntd:
          type: number
          minimum: 0.0
          title: Recommended Limit Ntd
          description: 'Suggested credit limit in NT dollars: raise on APPROVE, cut
            on REVIEW, 0 on DECLINE (available limit frozen, outstanding balance still
            due)'
        top_risk_factors:
          items:
            type: string
          type: array
          title: Top Risk Factors
          description: Human-readable drivers of the score
        policy_guardrails:
          additionalProperties:
            type: string
          type: object
          title: Policy Guardrails
          description: Deterministic compliance checks
        request_id:
          type: string
          title: Request Id
        latency_ms:
          type: number
          title: Latency Ms
      type: object
      required:
      - model_version
      - served_by
      - default_prediction
      - default_probability
      - credit_score
      - credit_tier
      - risk_decision
      - recommended_limit_ntd
      - request_id
      - latency_ms
      title: PredictionResponse
      description: Scoring result for ``POST /api/v1/predict``.
      examples:
      - credit_score: 420
        credit_tier: HIGH_RISK
        default_prediction: 1
        default_probability: 0.7812
        latency_ms: 11.4
        model_version: '3'
        policy_guardrails:
          age_verification: PASS
          delinquency_guardrail: ELEVATED_DEFAULT_RISK
          utilization_ceiling_check: OVER_UTILIZED_WARNING
        recommended_limit_ntd: 0.0
        request_id: req_4f1c2b7e9a0d4c3b8e2f6a1d0c9b8a7e
        risk_decision: DECLINE
        served_by: mlflow_registry
        top_risk_factors:
        - 'Severe Delinquency: PAY_0=2 indicates 2+ months payment default'
        - 'Excessive Credit Line Utilization: 97.5% of limit consumed'
        - 'Demographic Cohort: Young cardholder profile (23yo) with short account
          history'
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      description: Static API key issued to each client. Configure server keys with
        the `API_KEYS` env var.
      in: header
      name: X-API-Key
...